garlochi

+++ PRACTICA EXAMEN/IGProjects_x64_VS2022/IG1App/Camera.h
@@ -66,9 +66,9 @@ public:
 	void upload() const;
 
 protected:
-	glm::dvec3 mEye = {0.0, 0.0, 500.0}; // camera's position
-	glm::dvec3 mLook = {0.0, 0.0, 0.0};  // target's position
-	glm::dvec3 mUp = {0.0, 1.0, 0.0};    // the up vector
+	glm::vec3 mEye = {0.0, 0.0, 500.0}; // camera's position
+	glm::vec3 mLook = {0.0, 0.0, 0.0};  // target's position
+	glm::vec3 mUp = {0.0, 1.0, 0.0};    // the up vector
 
 	// Atributo de la clase Camera
 	// Ejes de la cámara ap 38.
